test_task/requirements.txt
asgiref==3.7.2
Django==4.2.9
django-pandas==0.6.6
djangorestframework==3.14.0
et-xmlfile==1.1.0
numpy==1.26.3
openpyxl==3.1.2
pandas==2.1.4
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
gunicorn==21.2.0
